speculative_decoding/requirements.txt
torch>=2.0.0
transformers>=4.40.0
accelerate>=0.30.0
safetensors>=0.4.0
tqdm>=4.65.0
numpy>=1.24.0
coloredlogs>=15.0
pyyaml>=6.0
datasets>=2.0.0
scikit-learn>=1.0.0
tiktoken>=0.7.0
protobuf>=4.0.0 